nn_inference/requirements.txt
absl-py==0.6.0
astor==0.7.1
astroid==2.1.0
bleach==1.5.0
certifi==2018.10.15
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29
dask==0.20.2
decorator==4.3.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
html5lib==0.9999999
imutils==0.5.2
isort==4.3.4
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keras-resnet==0.1.0
keras-retinanet==0.0.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
matplotlib==3.0.1
mccabe==0.6.1
networkx==1.11
numpy==1.15.3
opencv-contrib-python==3.4.4.19
opencv-python==3.4.3.18
pandas==0.24.0
pgmpy==0.1.7
Pillow==5.3.0
protobuf==3.6.1
pycocotools==2.0
pylint==2.2.2
pyparsing==2.2.2
python-dateutil==2.7.4
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.1
scipy==1.2.0
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
toolz==0.9.0
typed-ast==1.2.0
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.11.1
